cadastro categoria fornecedor - secretaria parcialmente pronto

+++ sistema/app/Http/Controllers/FornecedorController.php
@@ -10,11 +10,16 @@ use App\Http\Controllers\Controller;
 use App\Fornecedor;
 use Amranidev\Ajaxis\Ajaxis;
 use URL;
+use Gate;
 
 class FornecedorController extends Controller
 {
 
-     /**
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -22,17 +27,25 @@ class FornecedorController extends Controller
     public function index()
     {
         try {
-            $titulo = "Fornecedores ";
-            $fornecedors = Fornecedor::paginate(20);
-            return view('fornecedor.index', compact('fornecedors', 'titulo'));
+            if (Gate::allows('View_fornecedor')) {
+                $titulo = "Fornecedores ";
+                $fornecedors = Fornecedor::paginate(20);
+                return view('fornecedor.index', compact('fornecedors', 'titulo'));
+            }else{
+                return view('errors.sem_permissao');
+            }
         } catch (\Throwable $th) {
             dd($th);
         }
     }
     public function cadastrar()
     {
-        $titulo = "Fornecedores ";
-        return view('fornecedor.formulario', compact('titulo')) ;
+        if (Gate::allows('Insert_fornecedor')) {
+            $titulo = "Fornecedores ";
+            return view('fornecedor.formulario', compact('titulo'));
+        } else {
+            return view('errors.sem_permissao');
+        }
     }
     /**
      * Show the form for creating a new resource.
@@ -41,8 +54,12 @@ class FornecedorController extends Controller
      */
     public function insert(Request $request)
     {
-        Fornecedor::create($request->all());
-        return redirect('fornecedor');
+        if (Gate::allows('Insert_fornecedor')) {
+            Fornecedor::create($request->all());
+            return redirect('fornecedor');
+        } else {
+            return view('errors.sem_permissao');
+        }
     }
 
 
@@ -54,9 +71,13 @@ class FornecedorController extends Controller
      */
     public function editar(Fornecedor $fornecedor)
     {
-        $fornecedor = Fornecedor::findOrFail($fornecedor->id);
-        $titulo = "Editar ";
-        return view('fornecedor.formulario', compact('fornecedor', 'titulo'));
+        if (Gate::allows('Edit_fornecedor')) {
+            $fornecedor = Fornecedor::findOrFail($fornecedor->id);
+            $titulo = "Editar ";
+            return view('fornecedor.formulario', compact('fornecedor', 'titulo'));
+        } else {
+            return view('errors.sem_permissao');
+        }
     }
 
     /**
@@ -68,11 +89,14 @@ class FornecedorController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $fornecedor = Fornecedor::findOrFail($id);
-        $formRequest = $request->all();
-        $update =  $fornecedor->update($formRequest);
-
-        return redirect('fornecedor');
+        if (Gate::allows('Edit_fornecedor')) {
+            $fornecedor = Fornecedor::findOrFail($id);
+            $formRequest = $request->all();
+            $update =  $fornecedor->update($formRequest);
+            return redirect('fornecedor');
+        } else {
+            return view('errors.sem_permissao');
+        }
     }
 
     /**
@@ -83,8 +107,30 @@ class FornecedorController extends Controller
      */
     public function deletar($id)
     {
-        $fornecedor = Fornecedor::findOrFail($id);
-        $fornecedor->delete();
-        return redirect('frnecedor');
+        if (Gate::allows('Delete_fornecedor')) {
+            $fornecedor = Fornecedor::findOrFail($id);
+            $fornecedor->delete();
+            return redirect('fornecedor');
+        } else {
+            return view('errors.sem_permissao');
+        }
+    }
+
+    public function search(Request $request)
+    {
+        try {
+            $titulo = 'Pesquisa de fornecedores com o nome de  ' . $request->get('table_search');
+            if (Gate::allows('View_fornecedor')) {
+                $fornecedors = new Fornecedor();
+                $search = $request->get('table_search');
+                $fornecedors = Fornecedor::where('nome_fornecedor', 'like', '%' . $search . '%')->paginate(10);
+
+                return view('fornecedor.index', compact('titulo', 'fornecedors'));
+            } else {
+                return view('errors.404');
+            }
+        } catch (\Throwable $th) {
+            return view('errors.404', $th);
+        }
     }
 }
